:package & :wrench: change project structure & gitignore
