+++ .gitignore
@@ -5,6 +5,11 @@
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio, WebStorm and Rider
 # Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
 
+resource/database.properties
+.idea
+.idea/*.xml
+*.iml
+
 # User-specific stuff
 .idea/**/workspace.xml
 .idea/**/tasks.xml
@@ -148,9 +153,6 @@ replay_pid*
 
 # Icon must end with two \r
 Icon
-
-resource/database.properties
-WBS_br_31.iml
 
 # Thumbnails
 ._*
