--- src/main/java/com/ssg/Main.java
@@ -1,14 +1,0 @@
-package main.java.com.ssg;
-
-import main.java.com.ssg.library.Script;
-
-public class Main {
-
-  public static Script  script = new Script();
-
-  public static void main(String[] args) {
-    script.printStartMenu();
-    System.out.println();
-;
-  }
-}
