--- src/main/java/com/ssg/authentication/Authentication.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.authentication;
-
-public class Authentication {
-
-}
