--- src/main/java/com/ssg/controller/inbound/Inbound.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.controller.inbound;
-
-public class Inbound {
-
-}
