--- src/main/java/com/ssg/dao/inbound/InboundDAO.java
@@ -1,7 +1,0 @@
-package main.java.com.ssg.dao.inbound;
-
-public class InboundDAO {
-   public int multiply(int x, int y){
-   return x * y;
-  }
-}
