--- src/main/java/com/ssg/dao/inbound/ProductDAO.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.dao.inbound;
-
-public class ProductDAO {
-
-}
