--- src/main/java/com/ssg/dao/inbound/WarehouseDAO.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.dao.inbound;
-
-public class WarehouseDAO {
-
-}
