--- src/main/java/com/ssg/dao/inbound/stock/StockDAO.java
@@ -1,6 +1,0 @@
-package main.java.com.ssg.dao.inbound.stock;
-
-public class StockDAO {
-
-
-}
