--- src/main/java/com/ssg/dto/User.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.dto;
-
-public class User {
-
-}
