--- src/main/java/com/ssg/exception/Exception.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.exception;
-
-public class Exception {
-
-}
