--- src/main/java/com/ssg/library/Library.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.library;
-
-public class Library {
-
-}
