--- src/main/java/com/ssg/permission/Permission.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.permission;
-
-public class Permission {
-
-}
