--- src/main/java/com/ssg/service/Service.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.service;
-
-public class Service {
-
-}
