--- src/main/java/com/ssg/utility/Utility.java
@@ -1,5 +1,0 @@
-package main.java.com.ssg.utility;
-
-public class Utility {
-
-}
