+++ src/ssg/Main.java
@@ -1,0 +1,14 @@
+package ssg;
+
+import ssg.library.Script;
+
+public class Main {
+
+  public static Script script = new Script();
+
+  public static void main(String[] args) {
+    script.printStartMenu();
+    System.out.println();
+;
+  }
+}
