+++ src/ssg/authentication/Authentication.java
@@ -1,0 +1,5 @@
+package ssg.authentication;
+
+public class Authentication {
+
+}
