+++ src/ssg/controller/inbound/Inbound.java
@@ -1,0 +1,5 @@
+package ssg.controller.inbound;
+
+public class Inbound {
+
+}
