+++ src/ssg/dao/inbound/InboundDAO.java
@@ -1,0 +1,7 @@
+package ssg.dao.inbound;
+
+public class InboundDAO {
+   public int multiply(int x, int y){
+   return x * y;
+  }
+}
