+++ src/ssg/dao/inbound/ProductDAO.java
@@ -1,0 +1,5 @@
+package ssg.dao.inbound;
+
+public class ProductDAO {
+
+}
