+++ src/ssg/dao/inbound/WarehouseDAO.java
@@ -1,0 +1,5 @@
+package ssg.dao.inbound;
+
+public class WarehouseDAO {
+
+}
