+++ src/ssg/dao/inbound/stock/StockDAO.java
@@ -1,0 +1,6 @@
+package ssg.dao.inbound.stock;
+
+public class StockDAO {
+
+
+}
