+++ src/ssg/dto/User.java
@@ -1,0 +1,5 @@
+package ssg.dto;
+
+public class User {
+
+}
