+++ src/ssg/exception/Exception.java
@@ -1,0 +1,5 @@
+package ssg.exception;
+
+public class Exception {
+
+}
