+++ src/ssg/library/Library.java
@@ -1,0 +1,5 @@
+package ssg.library;
+
+public class Library {
+
+}
