+++ src/ssg/permission/Permission.java
@@ -1,0 +1,5 @@
+package ssg.permission;
+
+public class Permission {
+
+}
