+++ src/ssg/service/Service.java
@@ -1,0 +1,5 @@
+package ssg.service;
+
+public class Service {
+
+}
