+++ src/ssg/utility/Utility.java
@@ -1,0 +1,5 @@
+package ssg.utility;
+
+public class Utility {
+
+}
